+++ tests/tests.py
@@ -10,7 +10,6 @@ import pandas as pd
 import pyarrow as pa
 import pyogrio
 import pytest
-import rasterio
 from shapely.geometry import LineString, MultiLineString, Point
 
 import ouroboros as ob
@@ -723,10 +722,6 @@ class TestUsage:
 
 class TestRaster:
     def test_raster_to_tif(self, tmp_path, gdb_data_test):
-        if (
-            "gdb" not in rasterio.drivers.raster_driver_extensions()
-        ):  # TODO get this working on macOS and Linux test runners
-            assert False
 
         ob.raster_to_tif(
             gdb_path=gdb_data_test,
@@ -748,8 +743,6 @@ class TestRaster:
             tif_path=str(tif_path),
             write_kwargs={"tiled": True},
         )
-
-        # assert tif_path.exists()
 
     def test_tif_to_raster(self):
         with pytest.raises(NotImplementedError):